refactor: promote server.rs to server/ module directory
Split SSH key management (generate, import, use, unset, get_public_key)
into server/keys.rs. Entity struct, ConfigEntity impl, CRUD macro, and
helpers remain in server/mod.rs. Public API unchanged via pub use keys::*.
