+++ src/core/server/keys.rs
@@ -1,91 +1,9 @@
-use crate::config::{self, ConfigEntity};
 use crate::error::{Error, Result};
 use crate::local_files::{self, FileSystem};
-use crate::output::{CreateOutput, MergeOutput, RemoveResult};
-use crate::paths;
-use crate::project;
-use serde::{Deserialize, Serialize};
+use serde::Serialize;
 use std::process::Command;
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-
-pub struct Server {
-    #[serde(skip_deserializing, default)]
-    pub id: String,
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub aliases: Vec<String>,
-    pub host: String,
-    pub user: String,
-    #[serde(default = "default_port")]
-    pub port: u16,
-    #[serde(default)]
-    pub identity_file: Option<String>,
-}
-
-fn default_port() -> u16 {
-    22
-}
-
-impl Server {
-    pub fn is_valid(&self) -> bool {
-        !self.host.is_empty() && !self.user.is_empty()
-    }
-}
-
-impl ConfigEntity for Server {
-    const ENTITY_TYPE: &'static str = "server";
-    const DIR_NAME: &'static str = "servers";
-
-    fn id(&self) -> &str {
-        &self.id
-    }
-    fn set_id(&mut self, id: String) {
-        self.id = id;
-    }
-    fn not_found_error(id: String, suggestions: Vec<String>) -> Error {
-        Error::server_not_found(id, suggestions)
-    }
-    fn aliases(&self) -> &[String] {
-        &self.aliases
-    }
-    fn dependents(id: &str) -> Result<Vec<String>> {
-        let projects = project::list().unwrap_or_default();
-        Ok(projects
-            .iter()
-            .filter(|p| p.server_id.as_deref() == Some(id))
-            .map(|p| p.id.clone())
-            .collect())
-    }
-}
-
-// ============================================================================
-// Core CRUD - Generated by entity_crud! macro
-// ============================================================================
-
-entity_crud!(Server; merge);
-
-pub fn find_by_host(host: &str) -> Option<Server> {
-    list().ok()?.into_iter().find(|s| s.host == host)
-}
-
-pub fn key_path(id: &str) -> Result<std::path::PathBuf> {
-    paths::key(id)
-}
-
-// ============================================================================
-// Operations
-// ============================================================================
-
-pub fn set_identity_file(id: &str, identity_file: Option<String>) -> Result<Server> {
-    let mut server = load(id)?;
-    server.identity_file = identity_file;
-    save(&server)?;
-    Ok(server)
-}
-
-// ============================================================================
-// SSH Key Management
-// ============================================================================
+use super::{key_path, load, set_identity_file, Server};
 
 #[derive(Debug, Clone, Serialize)]
